requirements.txt
opencv-python>=4.8.1
numpy>=1.24.3
imagehash>=4.3.1
moviepy>=1.0.3
tqdm>=4.66.1
Pillow>=10.0.0
Flask>=3.0.2
Flask-WTF>=1.2.1
python-dotenv>=1.0.1
gunicorn>=21.2.0
gevent>=23.9.1
Werkzeug>=3.0.1 